refactor: move the code that changes the working directory to executed script, into 'linuxAIO'
Inside the 'installer-prep', the code will never change directories, as it's looking at the location of 'installer-prep', which will always be in the current directory. This is due to 'linxuAIO' not moving the working directory to the location of 'linuxAIO', if it was executed outside the correct directory. I just downloads 'installer-prep' into the current working dir.